EUI-284 test and refactor
