+++ src/users/store/effects/index.ts
@@ -1,6 +1,9 @@
 import { UsersEffects } from './users.effects';
+import { InviteUserEffects } from './invite-user.effects';
 
 export const effects: any[] = [
-  UsersEffects
+  UsersEffects,
+  InviteUserEffects
 ];
 export * from './users.effects';
+export * from './invite-user.effects';
