+++ src/users/store/effects/users.effects.spec.ts
@@ -5,10 +5,9 @@ import { of, throwError } from 'rxjs';
 import { provideMockActions } from '@ngrx/effects/testing';
 import * as fromUsersEffects from './users.effects';
 import { UsersEffects } from './users.effects';
-import { LoadUsersSuccess, LoadUsersFail, LoadUsers, SendInviteUser, InviteUserSuccess, InviteUserFail } from '../actions/user.actions';
+import { LoadUsersSuccess, LoadUsersFail, LoadUsers } from '../actions/user.actions';
+import { SendInviteUser, InviteUserSuccess, InviteUserFail } from '../actions/invite-user.actions';
 import { UsersService } from '../../services/users.service';
-import { InviteUserService } from '../../services/invite-user.service';
-import { AddFromData } from '../actions/invite-user.actions';
 
 describe('Users Effects', () => {
     let actions$;
@@ -17,9 +16,6 @@ describe('Users Effects', () => {
         'getListOfUsers',
     ]);
 
-    const InviteUsersServiceMock = jasmine.createSpyObj('InviteUserService', [
-        'inviteUser',
-    ]);
 
     beforeEach(() => {
         TestBed.configureTestingModule({
@@ -28,10 +24,6 @@ describe('Users Effects', () => {
                 {
                     provide: UsersService,
                     useValue: UsersServiceMock,
-                },
-                {
-                    provide: InviteUserService,
-                    useValue: InviteUsersServiceMock,
                 },
                 fromUsersEffects.UsersEffects,
                 provideMockActions(() => actions$)
@@ -64,41 +56,5 @@ describe('Users Effects', () => {
             expect(effects.loadUsers$).toBeObservable(expected);
         });
     });
-
-    describe('saveUsers$', () => {
-        it('should return a collection from user details - InviteUserSuccess', () => {
-            const payload = [{ payload: 'something' }];
-            InviteUsersServiceMock.inviteUser.and.returnValue(of(payload));
-            const requestPayload = {
-                firstName: 'Captain',
-                lastName: 'Caveman',
-                emailAddress: 'thecap@cave.com',
-                permissions: ['god']
-            };
-            const action = new SendInviteUser(requestPayload);
-            const completion = new InviteUserSuccess([{ payload: 'something' }]);
-            actions$ = hot('-a', { a: action });
-            const expected = cold('-b', { b: completion });
-            expect(effects.saveUsers$).toBeObservable(expected);
-        });
-    });
-
-    describe('saveUsers$ error', () => {
-        it('should return InviteUserFail', () => {
-            InviteUsersServiceMock.inviteUser.and.returnValue(throwError(new Error()));
-            const requestPayload = {
-                firstName: 'Captain',
-                lastName: 'Caveman',
-                emailAddress: 'thecap@cave.com',
-                permissions: ['god']
-            };
-            const action = new SendInviteUser(requestPayload);
-            const completion = new InviteUserFail(new Error());
-            actions$ = hot('-a', { a: action });
-            const expected = cold('-b', { b: completion });
-            expect(effects.saveUsers$).toBeObservable(expected);
-        });
-    });
-
 
 });
