+++ src/users/store/effects/users.effects.ts
@@ -1,10 +1,10 @@
-import {Injectable} from '@angular/core';
-import {Actions, Effect, ofType} from '@ngrx/effects';
+import { Injectable } from '@angular/core';
+import { Actions, Effect, ofType } from '@ngrx/effects';
 
 import * as usersActions from '../actions';
-import {catchError, map, switchMap} from 'rxjs/operators';
-import {of} from 'rxjs';
-import {InviteUserService, UsersService} from '../../services';
+import { catchError, map, switchMap } from 'rxjs/operators';
+import { of } from 'rxjs';
+import { UsersService } from '../../services';
 
 
 
@@ -12,9 +12,8 @@ import {InviteUserService, UsersService} from '../../services';
 export class UsersEffects {
   constructor(
     private actions$: Actions,
-    private usersService: UsersService,
-    private inviteUserSevice: InviteUserService
-  ) {}
+    private usersService: UsersService
+  ) { }
 
   @Effect()
   loadUsers$ = this.actions$.pipe(
@@ -23,18 +22,6 @@ export class UsersEffects {
       return this.usersService.getListOfUsers().pipe(
         map(userDetals => new usersActions.LoadUsersSuccess(userDetals)),
         catchError(error => of(new usersActions.LoadUsersFail(error)))
-      );
-    })
-  );
-
-  @Effect()
-  saveUsers$ = this.actions$.pipe(
-    ofType(usersActions.SEND_INVITE_USER),
-    map((action: usersActions.SendInviteUser) => action.payload),
-    switchMap((inviteUserFormData) => {
-      return this.inviteUserSevice.inviteUser(inviteUserFormData).pipe(
-        map(userDetails => new usersActions.InviteUserSuccess(userDetails)),
-        catchError(error => of(new usersActions.InviteUserFail(error)))
       );
     })
   );
